% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{argo_main}
\alias{argo_main}
\title{main function for argo}
\usage{
argo_main(save.folder = NULL)
}
\arguments{
\item{save.folder}{output folder to save graphics. If NULL then do not output graphics.}
}
\description{
main function that reproduce the results in ARGO paper
}
\examples{
\donttest{
argo_main()
}

}
