% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{argox_main}
\alias{argox_main}
\title{main function for argox}
\usage{
argox_main(
  gt.folder,
  ili.folder,
  population.file,
  gft.file,
  mix,
  save.folder = NULL,
  NCORES = 8
)
}
\arguments{
\item{gt.folder}{folder with Google Trends files, which should be thousands of csv file such as "US-MA_fever cough.csv" or "US-NY_cold or flu.csv"}

\item{ili.folder}{folder with ILINet data files: "ILINet_nat.csv" and "ILINet_regional.csv"}

\item{population.file}{file path to population csv file}

\item{gft.file}{file path to Google Flu Trends csv file}

\item{mix}{the weighted avarage mixing of raw state-level Google Trends data. Set to be 0 for stand-alone model. Set to be 1/3 for spatial-pooling model.}

\item{save.folder}{output folder to save graphics. If NULL then do not output graphics.}

\item{NCORES}{number of parallel cpu cores to be used.}
}
\description{
Main function that reproduce the results in ARGOX paper. The datasets are available at Harvard Dataverse <doi:10.7910/DVN/2IVDGK>.
}
\references{
Yang, S., Ning, S. & Kou, S.C. Use Internet search data to accurately track state level influenza epidemics. Sci Rep 11, 4023 (2021)
}
