% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataparsing.R
\name{gt.parser.pub.api}
\alias{gt.parser.pub.api}
\title{Parsing each Google Trends file downloaded from Google Trends API}
\usage{
gt.parser.pub.api(gt.folder, f)
}
\arguments{
\item{gt.folder}{folder that contains Google Trends file}

\item{f}{filename for Google Trends file}
}
\description{
Parsing each Google Trends file downloaded from Google Trends API
}
