% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataparsing.R
\name{gt.parser.pub.web}
\alias{gt.parser.pub.web}
\title{Parsing each Google Trends file downloaded from website}
\usage{
gt.parser.pub.web(gt.folder, f)
}
\arguments{
\item{gt.folder}{folder that contains Google Trends file}

\item{f}{filename for Google Trends file}
}
\description{
Parsing each Google Trends file downloaded from website
}
