% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{heatmap_cor}
\alias{heatmap_cor}
\title{Heatmap plot of correlation matrix}
\usage{
heatmap_cor(cor_heat, lim = 1)
}
\arguments{
\item{cor_heat}{The coefficient matrix to draw heatmap}

\item{lim}{the limit to truncate for large coefficients for better presentation}
}
\value{
a graph on the default plot window
}
\description{
Heatmap plot of correlation matrix
}
\examples{
cor_coef <- matrix(runif(100, -1, 1), ncol=10)
colnames(cor_coef) <- paste0("col", 1:10)
rownames(cor_coef) <- paste0("row", 1:10)
heatmap_cor(cor_coef)

}
