% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logit_inv}
\alias{logit_inv}
\title{inverse logit function}
\usage{
logit_inv(x)
}
\arguments{
\item{x}{numeric value for inverse logit transformation}
}
\description{
inverse logit function
}
\examples{
logit_inv(0)
}
