% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataparsing.R
\name{parse_gt_weekly}
\alias{parse_gt_weekly}
\title{Parsing of Google Trends data}
\usage{
parse_gt_weekly(folder)
}
\arguments{
\item{folder}{folder with weekly Google Trends file}
}
\description{
Parsing of Google Trends data
}
\examples{

\donttest{
download.file("https://scholar.harvard.edu/files/syang/files/gt2016-10-24.zip",
file.path(tempdir(), "gt2016-10-24.zip"))
unzip(file.path(tempdir(), "gt2016-10-24.zip"), exdir = tempdir())
gt.folder <- file.path(tempdir(), "2016-10-19")
parsed_data <- parse_gt_weekly(gt.folder)
}

}
\references{
Yang, S., Santillana, M., & Kou, S. C. (2015). Accurate estimation of influenza epidemics using Google search data via ARGO. Proceedings of the National Academy of Sciences, \href{https://dx.doi.org/10.1073/pnas.1515373112}{doi: 10.1073/pnas.1515373112}.
}
