% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataparsing.R
\name{parse_unrevised_ili}
\alias{parse_unrevised_ili}
\title{Parsing of unrevised ili from online source}
\usage{
parse_unrevised_ili(type = "extdata", ili.weighted = TRUE)
}
\arguments{
\item{type}{the type of data folder to parse}

\item{ili.weighted}{indicator to use weighted ILI or not}
}
\description{
Parsing of unrevised ili from online source
}
\examples{

\donttest{
parse_unrevised_ili()
}

}
\references{
Yang, S., Santillana, M., & Kou, S. C. (2015). Accurate estimation of influenza epidemics using Google search data via ARGO. Proceedings of the National Academy of Sciences. <doi:10.1073/pnas.1515373112>.
}
