% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructor.R
\name{argo}
\alias{argo}
\title{Construct ARGO object}
\usage{
argo(data, exogen = xts::xts(NULL), N_lag = 1:52, N_training = 104,
  alpha = 1, use_all_previous = FALSE)
}
\arguments{
\item{data}{response variable as xts, last element can be NA}

\item{exogen}{exogenous predictors, default is NULL}

\item{N_lag}{vector of the AR model lags used,
if NULL then no AR lags will be used}

\item{N_training}{number of training points, if use_all_previous is true,
this is the least number of training points required}

\item{alpha}{penalty between lasso and ridge, alpha=1 represents lasso,
alpha=0 represents ridge, alpha=NA represents no penalty}

\item{use_all_previous}{boolean variable indicating whether to use "all available
data" or "a sliding window" for training}
}
\value{
A list of following named objects
\itemize{
\item \code{pred} An xts object with the same index as input,
which contains historical nowcast estimation

\item \code{coef} A matrix contains historical coefficient values of the predictors.

\item \code{parm} Parameter values passed to argo function.
}
}
\description{
Wrapper for ARGO. The real work horse is glmnet package and/or linear model.
}
\details{
This function takes the time series and exogenous variables (optional) as
input, and produces out-of-sample prediction for each time point.
}
\references{
Yang, S., Santillana, M., & Kou, S. C. (2015). Accurate estimation of influenza epidemics using Google search data via ARGO. Proceedings of the National Academy of Sciences, \href{https://dx.doi.org/10.1073/pnas.1515373112}{doi: 10.1073/pnas.1515373112}.
}

