% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_argo}
\alias{summary_argo}
\title{performance summary of ARGO applied on CDC's ILI data}
\usage{
summary_argo(GFT_xts, model_names, legend_names, periods,
  whole_period = "2009-03/2015-10")
}
\value{
A list of summary tables for the inputed periods, including RMSE, MAE, MAPE, corr
}
\description{
This function is only used to reproduce the ARGO table of PNAS paper.
If you use this summary function for other dataset, an error is highly likely to occur.
}
\examples{
summary_argo
}

